requirements.txt
streamlit
groq
python-dotenv
pandas
gTTS
python-docx
pdfplumber
youtube-transcript-api
duckduckgo-search
yt-dlp